+++ CHANGELOG.markdown
@@ -4,5 +4,6 @@ html5media changelog
 HEAD
 ----
 
+*   Flowplayer now runs with wmode opaque by default.
 *   Updated Flowplayer binaries.
 *   Improved build process.